* Added Tag Barrack and Tag Main Base * Dynamic find barracks and main base * Added MULTIPLAYER (LockStep Model) * Added Network Actions * Fixed health visualization * Collect and unload only for owner * Two base on map * Small fixed balance:  ** Increased cost Soldier (100 -> 125)  ** Increased cost Worker (50 -> 75)  ** Increased time build Worker (1 sec -> 7 sec)

+++ Assets/Scripts/ResourcesController.cs
@@ -13,7 +13,7 @@ public class ResourcesController : MonoBehaviour {
         text = GameObject.Find("TextResourcesCount").GetComponent<UnityEngine.UI.Text>();
 
         resources = 200;
-        text.text = resources.ToString();
+        text.text = ((int)resources).ToString();
     }
 
     public float Add(float value)
@@ -27,7 +27,7 @@ public class ResourcesController : MonoBehaviour {
             added = resources - oldRes;
         }
 
-        text.text = resources.ToString();
+        text.text = ((int)resources).ToString();
 
         return added;
     }
@@ -37,7 +37,7 @@ public class ResourcesController : MonoBehaviour {
         resources -= value;
         if (resources < 0)
             resources = 0;
-        text.text = resources.ToString();
+        text.text = ((int)resources).ToString();
     }
 
     public bool isHave(float value)
